Cleaning and Updates consolidation of MQTTdata.py
