--- Communications/CSV/16072024/data_140445.csv
@@ -1,2 +1,0 @@
-Date,CPUTemp,CPU Usage,RAM Usage,Acceleration,Gyro,Magnetic,Pressure,BMP Temp,Altitude,UVA,UVB,UVC,UV Temp,Temperature
-14:04:50,60.05,0.5,14.1,"[0.1197100830078125, -1.898601916503906, -9.871293444824218]","[-0.007194487190839696, 0.007194487190839696, 0.002398162396946565]","[17.25, 71.85, -30.15]",Error,Error,Error,1.22955322265625,0.271087646484375,0.13677978515625,29.549999999999997,Error
