--- Communications/CSV/16072024/data_140750.csv
@@ -1,11 +1,0 @@
-Date,CPUTemp,CPU Usage,RAM Usage,Acceleration,Gyro,Magnetic,Pressure,BMP Temp,Altitude,UVA,UVB,UVC,UV Temp,Temperature
-14:07:52,61.7,Error,14.1,"[0.086191259765625, -1.8770541015625, -9.91438907470703]","[-0.006661562213740459, 0.009059724610687023, -0.000799387465648855]","[21.15, 71.25, -31.799999999999997]",Error,Error,Error,1.23992919921875,0.27685546875,0.139312744140625,29.549999999999997,Error
-14:07:56,61.15,Error,14.2,"[0.15322890625, -1.8698714965820311, -9.93354268798828]","[-0.006928024702290077, 0.0034640123511450386, 0.000799387465648855]","[21.15, 71.25, -31.349999999999998]",Error,Error,Error,1.20880126953125,0.248016357421875,0.11651611328125,29.5,Error
-14:08:00,61.7,Error,14.2,"[0.11252747802734374, -1.8171990600585937, -9.904812268066406]","[-0.007460949679389314, 0.0002664624885496183, -0.0021316999083969466]","[21.9, 71.55, -32.1]",Error,Error,Error,1.2347412109375,0.26531982421875,0.134246826171875,29.549999999999997,Error
-14:08:03,62.25,1.2,14.2,"[0.07900865478515624, -1.7980454467773437, -9.888052856445311]","[-0.008526799633587786, 0.007727412167938932, -0.0026646248854961837]","[22.349999999999998, 71.55, -31.799999999999997]",Error,Error,Error,1.22955322265625,0.2537841796875,0.12664794921875,29.549999999999997,Error
-14:08:07,62.25,0.5,14.2,"[0.16519991455078123, -1.8698714965820311, -9.86411083984375]","[-0.007194487190839696, 0.0061286372366412215, 0.002398162396946565]","[22.65, 71.25, -32.1]",Error,Error,Error,1.2139892578125,0.248016357421875,0.119049072265625,29.549999999999997,Error
-14:08:11,62.25,Error,14.2,"[0.11013327636718749, -1.8650830932617186, -9.87608184814453]","[-0.0130566619389313, 0.008793262122137404, -0.0010658499541984733]","[21.9, 72.0, -31.799999999999997]",Error,Error,Error,1.2347412109375,0.259552001953125,0.1317138671875,29.549999999999997,Error
-14:08:18,62.8,14.6,14.2,"[0.11731588134765625, -1.8650830932617186, -9.8449572265625]","[-0.004529862305343512, 0.011724349496183208, 0.0013323124427480919]","[22.5, 70.2, -32.699999999999996]",Error,Error,Error,1.26068115234375,0.322998046875,0.174774169921875,29.549999999999997,Error
-14:08:22,61.7,0.3,14.2,"[0.14604630126953125, -1.8124106567382812, -9.854534033203125]","[-0.010125574564885496, 0.0034640123511450386, 0.0013323124427480919]","[20.55, 71.85, -32.4]",Error,Error,Error,1.20361328125,0.236480712890625,0.1114501953125,29.549999999999997,Error
-14:08:26,62.25,Error,14.2,"[0.08140285644531249, -1.8435352783203125, -9.892841259765625]","[-0.012523736961832063, 0.004263399816793893, -0.00319754986259542]","[23.25, 72.14999999999999, -32.4]",Error,Error,Error,1.2554931640625,0.27685546875,0.144378662109375,29.5,Error
-14:08:30,61.7,14.3,14.2,"[0.143652099609375, -1.848323681640625, -9.89523546142578]","[-0.012523736961832063, 0.010392037053435116, -0.0018652374198473284]","[22.2, 71.85, -31.95]",Error,Error,Error,1.2347412109375,0.2999267578125,0.14691162109375,29.599999999999994,Error
