--- Communications/CSV/16072024/data_141558.csv
@@ -1,2 +1,0 @@
-Date,CPUTemp,CPU Usage,RAM Usage,Latitude,Longitude,Acceleration,Gyro,Magnetic,Pressure,BMP Temp,Altitude,UVA,UVB,UVC,UV Temp,Temperature
-14:16:02,62.8,25.1,14.1,Error,Error,"[0.03591302490234375, -1.8746598999023436, -9.859322436523437]","[-0.008526799633587786, 0.00799387465648855, 0.0]","[22.65, 71.25, -31.799999999999997]",Error,Error,Error,1.2762451171875,0.2537841796875,0.12158203125,29.799999999999997,Error
