--- Communications/CSV/16072024/data_232418.csv
@@ -1,5 +1,0 @@
-Date,CPUTemp,CPU Usage,RAM Usage,Latitude,Longitude,Altitude,Satellites,Elevation,Azimuth,UTC Time,Acceleration,Gyro,Magnetic,Pressure,BMP Temp,BMP Altitude,UVA,UVB,UVC,UV Temp,Temperature
-23:24:21,44.1,0.3,13.2,Error,Error,Error,Error,Error,Error,212419.00,"[0.5386953735351562, -1.7549498168945312, -9.9215716796875]","[-0.009059724610687023, 0.011990811984732825, 0.000799387465648855]","[-1.2, 56.1, -40.5]",Error,Error,Error,0.0518798828125,0.051910400390625,0.04052734375,24.799999999999997,Error
-23:24:24,43.0,0.5,13.2,Error,Error,Error,07,Error,Error,Error,"[0.47644613037109373, -1.0486603271484374, -9.442731347656249]","[0.03863706083969466, 0.04636447300763359, 0.05755589752671757]","[-2.4, 56.1, -41.25]",Error,Error,Error,0.062255859375,0.063446044921875,0.04559326171875,24.799999999999997,Error
-23:24:28,43.55,26.7,13.0,Error,Error,Error,07,Error,Error,Error,"[0.9504980590820312, -1.9823989746093749, -9.988609326171874]","[-0.07514242177099237, -0.0711454844427481, 0.09779173329770993]","[-0.44999999999999996, 58.8, -44.55]",Error,Error,Error,0.12969970703125,0.138427734375,0.12664794921875,24.75,Error
-23:24:32,44.65,25.1,13.0,Error,Error,Error,07,Error,Error,Error,"[0.6512228515624999, -1.9512743530273435, -9.852139831542969]","[-0.004263399816793893, 0.012523736961832063, 0.0013323124427480919]","[0.44999999999999996, 53.25, -37.199999999999996]",Error,Error,Error,0.10894775390625,0.121124267578125,0.098785400390625,24.75,Error
