--- Communications/CSV/16072024/data_232436.csv
@@ -1,8 +1,0 @@
-Date,CPUTemp,CPU Usage,RAM Usage,Latitude,Longitude,Altitude,Satellites,Elevation,Azimuth,UTC Time,Acceleration,Gyro,Magnetic,Pressure,BMP Temp,BMP Altitude,UVA,UVB,UVC,UV Temp,Temperature
-23:24:38,45.2,2.8,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.6177040283203125, -1.9009961181640624, -9.8449572265625]","[-0.008526799633587786, 0.01279019945038168, -0.0005329249770992366]","[-0.15, 52.5, -37.5]",Error,Error,Error,0.07781982421875,0.08074951171875,0.068389892578125,24.700000000000003,Error
-23:24:45,44.65,Error,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.6416460449218749, -1.9129671264648436, -9.811438403320311]","[-0.00639509972519084, 0.009326187099236643, 0.000799387465648855]","[0.0, 52.35, -37.35]",Error,Error,Error,0.062255859375,0.063446044921875,0.06585693359375,24.75,Error
-23:24:49,44.1,Error,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.6584054565429687, -1.9105729248046874, -9.823409411621093]","[-0.002398162396946565, 0.013323124427480917, 0.0021316999083969466]","[-1.5, 53.1, -36.15]",Error,Error,Error,0.07781982421875,0.08074951171875,0.07598876953125,24.650000000000006,Error
-23:24:53,44.65,2.0,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.5506663818359374, -1.7836802368164062, -10.165780249023436]","[0.025047473923664125, -0.058088822503816806, -0.012257274473282443]","[-3.15, 53.25, -36.0]",Error,Error,Error,0.0830078125,0.086517333984375,0.058258056640625,24.700000000000003,Error
-23:24:57,44.1,1.3,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.5889736083984375, -1.7645266235351562, -9.926360083007811]","[-0.008793262122137404, 0.032774886091603055, 0.0002664624885496183]","[-1.95, 53.25, -35.699999999999996]",Error,Error,Error,0.07781982421875,0.086517333984375,0.0709228515625,24.700000000000003,Error
-23:25:00,43.55,Error,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.7326257080078125, -1.7836802368164062, -9.873687646484374]","[-0.0130566619389313, 0.0034640123511450386, 0.0021316999083969466]","[-3.3, 52.35, -35.85]",Error,Error,Error,0.06744384765625,0.0692138671875,0.060791015625,24.650000000000006,Error
-23:25:04,45.2,5.5,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.6631938598632812, -1.7956512451171873, -9.94072529296875]","[-0.003996937328244275, 0.007460949679389314, 0.0021316999083969466]","[-2.4, 53.699999999999996, -36.0]",Error,Error,Error,0.10894775390625,0.121124267578125,0.113983154296875,24.650000000000006,Error
