--- Communications/CSV/16072024/data_233000.csv
@@ -1,21 +1,0 @@
-Date,CPUTemp,CPU Usage,RAM Usage,Latitude,Longitude,Altitude,Satellites,Elevation,Azimuth,UTC Time,Acceleration,Gyro,Magnetic,Pressure,BMP Temp,BMP Altitude,UVA,UVB,UVC,UV Temp,Temperature
-23:30:03,46.3,0.3,13.0,Error,Error,Error,Error,Error,Error,213001.00,"[0.6655880615234374, -1.7717092285156248, -9.849745629882811]","[-0.009326187099236643, 0.009326187099236643, 0.0005329249770992366]","[-2.55, 54.3, -36.9]",Error,Error,Error,0.01556396484375,0.01153564453125,0.02532958984375,24.200000000000003,Error
-23:30:06,45.75,0.2,13.1,Error,Error,Error,07,Error,Error,Error,"[0.6775590698242188, -1.7717092285156248, -9.900023864746093]","[-0.010924962030534353, 0.00159877493129771, -0.00639509972519084]","[-2.25, 52.65, -35.85]",Error,Error,Error,0.06744384765625,0.074981689453125,0.058258056640625,24.150000000000006,Error
-23:30:10,46.3,Error,13.1,Error,Error,Error,07,Error,Error,Error,"[0.6440402465820312, -1.7884686401367187, -9.832986218261718]","[-0.008526799633587786, 0.006661562213740459, 0.0005329249770992366]","[-3.9, 54.0, -36.0]",Error,Error,Error,0.09857177734375,0.10382080078125,0.08612060546875,24.150000000000006,Error
-23:30:17,45.75,3.5,13.1,Error,Error,Error,07,Error,Error,Error,"[0.6464344482421874, -1.8052280517578123, -9.87608184814453]","[-0.004263399816793893, 0.010924962030534353, 0.0061286372366412215]","[-3.5999999999999996, 54.0, -35.699999999999996]",Error,Error,Error,0.0830078125,0.086517333984375,0.0810546875,24.150000000000006,Error
-23:30:24,45.75,1.0,13.1,Error,Error,Error,07,Error,Error,Error,"[0.699106884765625, -1.7645266235351562, -9.828197814941406]","[-0.005862174748091604, 0.013323124427480917, 0.0005329249770992366]","[-4.05, 53.55, -35.55]",Error,Error,Error,0.06744384765625,0.0692138671875,0.060791015625,24.200000000000003,Error
-23:30:28,45.75,14.6,13.1,Error,Error,Error,07,Error,Error,Error,"[0.6871358764648438, -1.7908628417968748, -9.878476049804688]","[-0.00799387465648855, 0.007460949679389314, -0.0021316999083969466]","[-4.8, 53.55, -37.35]",Error,Error,Error,0.1141357421875,0.121124267578125,0.0911865234375,24.200000000000003,Error
-23:30:32,45.75,Error,13.1,Error,Error,Error,07,Error,Error,Error,"[0.6679822631835937, -1.7405846069335937, -9.861716638183593]","[-0.0021316999083969466, 0.007460949679389314, 0.000799387465648855]","[-4.05, 54.3, -36.15]",Error,Error,Error,0.08819580078125,0.09228515625,0.08612060546875,24.150000000000006,Error
-23:30:36,45.75,Error,13.1,Error,Error,Error,07,Error,Error,Error,"[0.6871358764648438, -1.7908628417968748, -9.885658654785155]","[-0.006928024702290077, 0.013323124427480917, 0.0018652374198473284]","[-3.9, 53.25, -36.75]",Error,Error,Error,0.04669189453125,0.046142578125,0.0506591796875,24.150000000000006,Error
-23:30:40,45.75,15.6,13.1,Error,Error,Error,07,Error,Error,Error,"[0.6679822631835937, -1.8171990600585937, -9.871293444824218]","[-0.007194487190839696, 0.01119142451908397, 0.0034640123511450386]","[-3.75, 53.55, -34.35]",Error,Error,Error,0.08819580078125,0.098052978515625,0.0709228515625,24.150000000000006,Error
-23:30:44,47.4,0.5,13.1,Error,Error,Error,07,Error,Error,Error,"[0.5913678100585937, -1.7405846069335937, -9.830592016601562]","[-0.010658499541984735, 0.005062787282442748, 0.00479632479389313]","[-3.75, 53.1, -36.15]",Error,Error,Error,0.06744384765625,0.0692138671875,0.068389892578125,24.150000000000006,Error
-23:30:48,45.75,Error,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.7206546997070312, -1.7741034301757812, -9.842563024902343]","[-0.002931087374045802, 0.0061286372366412215, -0.003730474839694657]","[-3.0, 54.3, -35.85]",Error,Error,Error,0.06744384765625,0.074981689453125,0.058258056640625,24.200000000000003,Error
-23:30:51,45.75,8.3,13.1,Error,Error,Error,07,Error,Error,Error,"[0.7110778930664062, -1.7597382202148437, -9.861716638183593]","[-0.00479632479389313, 0.01279019945038168, 0.002398162396946565]","[-3.3, 53.85, -36.15]",Error,Error,Error,0.02593994140625,0.0230712890625,0.03546142578125,24.200000000000003,Error
-23:30:55,45.75,Error,13.1,Error,Error,Error,07,Error,Error,Error,"[0.6847416748046875, -1.7334020019531249, -9.849745629882811]","[-0.01119142451908397, 0.01119142451908397, 0.00159877493129771]","[-3.75, 53.85, -35.85]",Error,Error,Error,0.03631591796875,0.040374755859375,0.04559326171875,24.150000000000006,Error
-23:30:59,44.65,0.5,13.1,Error,Error,Error,07,Error,Error,Error,"[0.6751648681640625, -1.7764976318359373, -9.897629663085937]","[-0.010392037053435116, 0.00799387465648855, 0.0002664624885496183]","[-3.75, 53.85, -37.35]",Error,Error,Error,0.04669189453125,0.051910400390625,0.04559326171875,24.150000000000006,Error
-23:31:03,45.75,2.0,13.1,Error,Error,Error,07,Error,Error,Error,"[0.689530078125, -1.8339584716796873, -9.873687646484374]","[-0.00639509972519084, 0.010392037053435116, -0.0010658499541984733]","[-3.15, 54.0, -38.25]",Error,Error,Error,0.0830078125,0.086517333984375,0.073455810546875,24.150000000000006,Error
-23:31:10,46.3,0.3,13.1,Error,Error,Error,07,Error,Error,Error,"[0.70868369140625, -1.8148048583984373, -9.847351428222655]","[-0.007727412167938932, 0.007460949679389314, -0.0002664624885496183]","[-3.3, 54.3, -36.15]",Error,Error,Error,0.08819580078125,0.09228515625,0.07598876953125,24.150000000000006,Error
-23:31:14,45.75,1.5,13.1,Error,Error,Error,07,Error,Error,Error,"[0.6296750366210937, -1.7764976318359373, -9.81862100830078]","[-0.01279019945038168, 0.006661562213740459, 0.002398162396946565]","[-3.15, 53.25, -35.699999999999996]",Error,Error,Error,0.07781982421875,0.08074951171875,0.073455810546875,24.150000000000006,Error
-23:31:21,45.75,0.5,13.1,Error,Error,Error,07,Error,Error,Error,"[0.6272808349609375, -1.7980454467773437, -9.871293444824218]","[-0.006661562213740459, 0.012523736961832063, 0.0026646248854961837]","[-4.05, 53.1, -36.6]",Error,Error,Error,0.062255859375,0.063446044921875,0.05572509765625,24.099999999999994,Error
-23:31:25,45.75,0.3,13.1,Error,Error,Error,07,Error,Error,Error,"[0.6919242797851562, -1.7764976318359373, -9.873687646484374]","[-0.01279019945038168, 0.007727412167938932, -0.002398162396946565]","[-3.15, 52.05, -35.25]",Error,Error,Error,0.1141357421875,0.12689208984375,0.108917236328125,24.099999999999994,Error
-23:31:29,45.75,15.8,13.1,Error,Error,Error,07,Error,Error,Error,"[0.6392518432617187, -1.7908628417968748, -9.897629663085937]","[-0.008526799633587786, 0.007194487190839696, 0.0002664624885496183]","[-1.7999999999999998, 53.85, -35.1]",Error,Error,Error,0.10894775390625,0.121124267578125,0.108917236328125,24.099999999999994,Error
