--- Communications/CSV/16072024/data_233335.csv
@@ -1,6 +1,0 @@
-Date,CPUTemp,CPU Usage,RAM Usage,Latitude,Longitude,Altitude,Satellites,Elevation,Azimuth,UTC Time,Acceleration,Gyro,Magnetic,Pressure,BMP Temp,BMP Altitude,UVA,UVB,UVC,UV Temp,Temperature
-23:33:38,46.3,0.5,13.0,Error,Error,Error,Error,Error,Error,213336.00,"[0.6607996582031249, -1.7884686401367187, -9.842563024902343]","[-0.006661562213740459, 0.007727412167938932, 0.002931087374045802]","[-3.15, 53.699999999999996, -36.449999999999996]",Error,Error,Error,0.03631591796875,0.03460693359375,0.04559326171875,24.049999999999997,Error
-23:33:41,45.75,11.1,13.0,Error,Error,Error,07,Error,Error,Error,"[0.6512228515624999, -1.7693150268554687, -9.904812268066406]","[-0.0053292497709923675, 0.008260337145038169, -0.0005329249770992366]","[-3.15, 52.949999999999996, -36.449999999999996]",Error,Error,Error,0.11932373046875,0.12689208984375,0.1114501953125,24.049999999999997,Error
-23:33:45,46.3,2.7,13.0,Error,Error,Error,07,Error,Error,Error,"[0.6727706665039062, -1.8052280517578123, -9.859322436523437]","[-0.010125574564885496, 0.006928024702290077, 0.010924962030534353]","[-4.35, 54.449999999999996, -34.5]",Error,Error,Error,0.06744384765625,0.0692138671875,0.0709228515625,24.049999999999997,Error
-23:33:49,45.75,25.1,13.1,Error,Error,Error,07,Error,Error,Error,"[0.6153098266601562, -1.9033903198242186, -9.868899243164062]","[-0.002398162396946565, 0.015188361847328247, 0.002398162396946565]","[-2.55, 55.8, -37.65]",Error,Error,Error,0.0933837890625,0.098052978515625,0.083587646484375,24.049999999999997,Error
-23:33:53,46.85,39.9,13.1,Error,Error,Error,Error,Error,Error,213342.00,"[0.70868369140625, -1.8555062866210936, -9.823409411621093]","[-0.010658499541984735, 0.008793262122137404, -0.0013323124427480919]","[-1.95, 54.0, -37.199999999999996]",Error,Error,Error,0.10894775390625,0.121124267578125,0.078521728515625,24.099999999999994,Error
