--- Communications/CSV/16072024/data_234046.csv
@@ -1,10 +1,0 @@
-Date,CPUTemp,CPU Usage,RAM Usage,Latitude,Longitude,Altitude,Satellites,Elevation,Azimuth,UTC Time,Acceleration,Gyro,Magnetic,Pressure,BMP Temp,BMP Altitude,UVA,UVB,UVC,UV Temp,Temperature
-23:40:47,48.5,Error,13.0,Error,Error,Error,Error,Error,Error,Error,"[0.737414111328125, -1.7669208251953124, -9.907206469726562]","[-0.0061286372366412215, 0.007194487190839696, 0.0005329249770992366]","[-4.6499999999999995, 54.0, -34.5]",Error,Error,Error,0.0726318359375,0.074981689453125,0.06585693359375,23.900000000000006,Error
-23:40:51,47.95,Error,13.0,Error,Error,Error,Error,Error,Error,Error,"[0.7781155395507812, -1.695094775390625, -9.871293444824218]","[-0.007460949679389314, 0.010392037053435116, -0.0026646248854961837]","[-4.5, 53.85, -35.1]",Error,Error,Error,0.02593994140625,0.0230712890625,0.032928466796875,23.950000000000003,Error
-23:40:55,48.5,1.3,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.8020575561523438, -1.7381904052734374, -9.82580361328125]","[0.015721286824427482, 0.01119142451908397, -0.0010658499541984733]","[-3.75, 54.3, -36.15]",Error,Error,Error,0.10894775390625,0.121124267578125,0.083587646484375,23.950000000000003,Error
-23:40:59,47.95,0.2,13.1,Error,Error,Error,00,Error,Error,Error,"[0.7134720947265625, -1.8076222534179687, -9.909600671386718]","[-0.00985911207633588, 0.014388974381679392, 0.00159877493129771]","[-4.05, 54.3, -35.1]",Error,Error,Error,0.04669189453125,0.046142578125,0.04052734375,23.950000000000003,Error
-23:41:04,47.4,0.5,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.7709329345703124, -1.8291700683593748, -10.007762939453125]","[-0.00479632479389313, 0.02105053659541985, 0.003730474839694657]","[-3.3, 55.05, -34.35]",Error,Error,Error,0.0726318359375,0.074981689453125,0.068389892578125,23.950000000000003,Error
-23:41:09,49.6,35.3,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.9337386474609375, -1.8411410766601561, -9.97184991455078]","[-0.009059724610687023, 0.011990811984732825, 0.003730474839694657]","[-2.25, 55.05, -35.55]",Error,Error,Error,0.103759765625,0.109588623046875,0.09625244140625,23.950000000000003,Error
-23:41:15,49.6,37.2,13.1,Error,Error,Error,Error,Error,Error,Error,"[1.0486603271484374, -1.8028338500976562, -9.598354455566406]","[0.010924962030534353, 0.03677182341984733, -0.05702297254961833]","[-1.2, 54.449999999999996, -34.5]",Error,Error,Error,0.103759765625,0.109588623046875,0.093719482421875,24.0,Error
-23:41:21,50.7,36.7,13.1,Error,Error,Error,Error,Error,Error,214118.00,"[0.7852981445312499, -1.7573440185546874, -9.830592016601562]","[-0.003730474839694657, 0.01119142451908397, -0.002931087374045802]","[-0.8999999999999999, 56.1, -34.949999999999996]",Error,Error,Error,0.12969970703125,0.138427734375,0.12158203125,23.900000000000006,Error
-23:41:27,50.15,39.7,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.7062894897460937, -1.7166425903320313, -9.911994873046874]","[-0.00959264958778626, 0.003730474839694657, -0.002931087374045802]","[-0.3, 55.8, -34.05]",Error,Error,Error,0.04669189453125,0.046142578125,0.0506591796875,23.950000000000003,Error
