--- Communications/CSV/16072024/data_234351.csv
@@ -1,16 +1,0 @@
-Date,CPUTemp,CPU Usage,RAM Usage,Latitude,Longitude,Altitude,Satellites,Elevation,Azimuth,UTC Time,Acceleration,Gyro,Magnetic,Pressure,BMP Temp,BMP Altitude,UVA,UVB,UVC,UV Temp,Temperature
-23:43:54,46.85,9.8,13.0,Error,Error,Error,Error,Error,Error,214352.00,"[0.5315127685546874, -2.099714855957031, -9.79946739501953]","[-0.007727412167938932, 0.005062787282442748, -0.0026646248854961837]","[68.1, 7.199999999999999, -98.55]",Error,Error,Error,0.06744384765625,0.063446044921875,0.063323974609375,24.150000000000006,Error
-23:43:57,46.85,0.5,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.5243301635742187, -2.04943662109375, -9.868899243164062]","[-0.00639509972519084, 0.0053292497709923675, 0.000799387465648855]","[68.1, 0.15, -93.75]",Error,Error,Error,0.04669189453125,0.03460693359375,0.03546142578125,24.200000000000003,Error
-23:44:01,47.95,0.2,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.6057330200195312, -1.9871873779296874, -9.854534033203125]","[-0.0061286372366412215, 0.002931087374045802, -0.0026646248854961837]","[68.7, 0.3, -95.1]",Error,Error,Error,0.062255859375,0.063446044921875,0.048126220703125,24.25,Error
-23:44:05,46.85,4.0,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.5698199951171875, -1.9919757812499999, -9.847351428222655]","[-0.010392037053435116, 0.011990811984732825, -0.0002664624885496183]","[69.14999999999999, 1.65, -92.7]",Error,Error,Error,0.07781982421875,0.08074951171875,0.068389892578125,24.25,Error
-23:44:09,46.85,0.5,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.5961562133789062, -1.9919757812499999, -9.87608184814453]","[-0.0026646248854961837, 0.009326187099236643, 0.0]","[66.0, 1.05, -93.6]",Error,Error,Error,0.0933837890625,0.098052978515625,0.073455810546875,24.25,Error
-23:44:16,47.4,Error,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.565031591796875, -2.032677209472656, -9.80665]","[-0.009059724610687023, 0.009059724610687023, 0.0013323124427480919]","[67.64999999999999, 0.15, -93.75]",Error,Error,Error,0.04669189453125,0.040374755859375,0.04559326171875,24.299999999999997,Error
-23:44:20,47.4,0.2,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.5291185668945312, -2.0087351928710935, -9.859322436523437]","[-0.012523736961832063, 0.007460949679389314, -0.00159877493129771]","[68.1, 1.3499999999999999, -94.5]",Error,Error,Error,0.10894775390625,0.1153564453125,0.09625244140625,24.299999999999997,Error
-23:44:27,47.4,0.5,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.5865794067382812, -2.0183119995117185, -9.81862100830078]","[-0.004263399816793893, 0.011457887007633588, -0.0026646248854961837]","[68.1, 1.3499999999999999, -93.75]",Error,Error,Error,0.0518798828125,0.046142578125,0.04052734375,24.349999999999994,Error
-23:44:31,47.95,15.0,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.5626373901367188, -2.032677209472656, -9.832986218261718]","[-0.009059724610687023, 0.008793262122137404, 0.0005329249770992366]","[68.25, 3.15, -93.6]",Error,Error,Error,0.12969970703125,0.132659912109375,0.113983154296875,24.349999999999994,Error
-23:44:35,46.85,0.5,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.5817910034179687, -1.9776105712890624, -9.82580361328125]","[-0.004529862305343512, 0.008260337145038169, 0.00479632479389313]","[69.14999999999999, -0.8999999999999999, -94.2]",Error,Error,Error,0.062255859375,0.05767822265625,0.058258056640625,24.450000000000003,Error
-23:44:39,47.4,1.5,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.5674257934570313, -2.0087351928710935, -9.775525378417969]","[-0.00985911207633588, 0.014921899358778627, -0.000799387465648855]","[67.35, 0.44999999999999996, -94.2]",Error,Error,Error,0.0726318359375,0.0692138671875,0.053192138671875,24.450000000000003,Error
-23:44:42,46.85,15.8,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.584185205078125, -2.023100402832031, -9.849745629882811]","[-0.00639509972519084, 0.010125574564885496, -0.0005329249770992366]","[68.7, 1.5, -94.35]",Error,Error,Error,0.12969970703125,0.132659912109375,0.119049072265625,24.5,Error
-23:44:46,46.85,0.3,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.612915625, -2.0278888061523435, -9.847351428222655]","[-0.005595712259541985, 0.009059724610687023, 0.004263399816793893]","[68.7, 2.25, -93.3]",Error,Error,Error,0.062255859375,0.063446044921875,0.058258056640625,24.599999999999994,Error
-23:44:50,47.4,Error,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.6296750366210937, -2.0183119995117185, -9.765948571777344]","[-0.009326187099236643, 0.006928024702290077, 0.0026646248854961837]","[67.95, 2.25, -93.3]",Error,Error,Error,0.05706787109375,0.051910400390625,0.03546142578125,24.650000000000006,Error
-23:44:54,47.4,13.6,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.6105214233398437, -1.9991583862304687, -9.880870251464843]","[-0.003996937328244275, 0.009326187099236643, -0.0021316999083969466]","[68.1, 0.8999999999999999, -93.45]",Error,Error,Error,0.05706787109375,0.051910400390625,0.053192138671875,24.700000000000003,Error
