--- Communications/CSV/16072024/data_234642.csv
@@ -1,22 +1,0 @@
-Date,CPUTemp,CPU Usage,RAM Usage,Latitude,Longitude,Altitude,Satellites,Elevation,Azimuth,UTC Time,Acceleration,Gyro,Magnetic,Pressure,BMP Temp,BMP Altitude,UVA,UVB,UVC,UV Temp,Temperature
-23:46:44,50.7,9.5,12.9,Error,Error,Error,Error,Error,Error,214642.00,"[0.5722141967773438, -1.9704279663085937, -9.849745629882811]","[-0.007460949679389314, 0.010392037053435116, 0.0]","[67.35, 1.3499999999999999, -94.5]",Error,Error,Error,0.12451171875,0.138427734375,0.10638427734375,25.549999999999997,Error
-23:46:47,50.15,Error,12.9,Error,Error,Error,Error,Error,Error,Error,"[0.7517793212890624, -2.0278888061523435, -9.79946739501953]","[0.003730474839694657, -0.008793262122137404, 0.007194487190839696]","[68.85, 3.3, -93.75]",Error,Error,Error,0.020751953125,0.028839111328125,0.012664794921875,25.599999999999994,Error
-23:46:51,50.15,0.7,12.9,Error,Error,Error,Error,Error,Error,Error,"[0.5889736083984375, -2.0183119995117185, -9.83777462158203]","[-0.005062787282442748, 0.0034640123511450386, 0.0010658499541984733]","[67.64999999999999, 1.3499999999999999, -94.5]",Error,Error,Error,0.03631591796875,0.040374755859375,0.022796630859375,25.599999999999994,Error
-23:46:55,50.15,14.6,13.0,Error,Error,Error,Error,Error,Error,Error,"[0.565031591796875, -1.917755529785156, -9.847351428222655]","[-0.004263399816793893, 0.009059724610687023, 0.0005329249770992366]","[67.35, 2.1, -93.75]",Error,Error,Error,0.103759765625,0.1153564453125,0.093719482421875,25.650000000000006,Error
-23:46:59,50.15,Error,13.0,Error,Error,Error,Error,Error,Error,Error,"[0.5674257934570313, -1.9201497314453124, -9.854534033203125]","[-0.004529862305343512, 0.00799387465648855, -0.0002664624885496183]","[68.39999999999999, 0.15, -94.95]",Error,Error,Error,0.0726318359375,0.08074951171875,0.0506591796875,25.650000000000006,Error
-23:47:03,50.15,0.7,13.0,Error,Error,Error,Error,Error,Error,Error,"[0.63206923828125, -1.9776105712890624, -9.816226806640625]","[-0.005062787282442748, 0.01119142451908397, 0.0026646248854961837]","[68.7, 0.75, -94.05]",Error,Error,Error,0.020751953125,0.0230712890625,0.012664794921875,25.700000000000003,Error
-23:47:07,50.15,14.6,13.0,Error,Error,Error,Error,Error,Error,Error,"[0.5985504150390625, -1.946485949707031, -9.842563024902343]","[-0.004263399816793893, 0.008793262122137404, -0.000799387465648855]","[67.64999999999999, 0.8999999999999999, -94.5]",Error,Error,Error,0.0933837890625,0.10382080078125,0.08612060546875,25.700000000000003,Error
-23:47:11,50.15,0.5,13.0,Error,Error,Error,12,Error,Error,Error,"[0.5674257934570313, -1.9776105712890624, -9.80665]","[-0.0053292497709923675, 0.011457887007633588, 0.0010658499541984733]","[68.85, 3.3, -93.75]",Error,Error,Error,0.02593994140625,0.028839111328125,0.020263671875,25.700000000000003,Error
-23:47:14,50.7,0.2,13.0,Error,Error,Error,Error,Error,Error,Error,"[0.5482721801757812, -1.9297265380859374, -9.873687646484374]","[-0.007194487190839696, 0.005595712259541985, -0.000799387465648855]","[68.1, 2.55, -94.95]",Error,Error,Error,0.03631591796875,0.040374755859375,0.032928466796875,25.700000000000003,Error
-23:47:18,50.7,12.8,13.0,Error,Error,Error,Error,Error,Error,Error,"[0.5674257934570313, -1.905784521484375, -9.794678991699218]","[-0.006928024702290077, 0.007460949679389314, -0.00319754986259542]","[68.85, 2.85, -95.25]",Error,Error,Error,0.05706787109375,0.063446044921875,0.043060302734375,25.75,Error
-23:47:25,50.15,0.5,13.0,Error,Error,Error,Error,Error,Error,Error,"[0.5698199951171875, -1.946485949707031, -9.78031378173828]","[-0.006661562213740459, 0.01279019945038168, 0.0021316999083969466]","[68.7, 0.75, -94.8]",Error,Error,Error,0.11932373046875,0.132659912109375,0.103851318359375,25.75,Error
-23:47:32,51.8,0.2,13.0,Error,Error,Error,Error,Error,Error,Error,"[0.5722141967773438, -2.023100402832031, -9.823409411621093]","[-0.007460949679389314, 0.009326187099236643, -0.000799387465648855]","[0.0, 0.0, 0.0]",Error,Error,Error,0.05706787109375,0.063446044921875,0.04052734375,25.799999999999997,Error
-23:47:39,51.25,0.5,13.0,Error,Error,Error,Error,Error,Error,Error,"[0.584185205078125, -1.9321207397460936, -9.80665]","[-0.005062787282442748, 0.008793262122137404, 0.0005329249770992366]","[70.35, 2.1, -95.25]",Error,Error,Error,0.12451171875,0.138427734375,0.108917236328125,25.849999999999994,Error
-23:47:43,51.25,14.6,13.0,Error,Error,Error,Error,Error,Error,Error,"[0.5985504150390625, -1.9393033447265624, -9.871293444824218]","[-0.00479632479389313, 0.012523736961832063, -0.0005329249770992366]","[68.1, 0.44999999999999996, -94.2]",Error,Error,Error,0.10894775390625,0.121124267578125,0.093719482421875,25.849999999999994,Error
-23:47:47,51.25,0.2,13.0,Error,Error,Error,Error,Error,Error,Error,"[0.622492431640625, -2.0063409912109376, -9.792284790039062]","[-0.011724349496183208, 0.004529862305343512, -0.0018652374198473284]","[69.14999999999999, 0.44999999999999996, -94.95]",Error,Error,Error,0.05706787109375,0.063446044921875,0.048126220703125,25.849999999999994,Error
-23:47:54,51.8,9.5,13.0,Error,Error,Error,Error,Error,Error,Error,"[0.5817910034179687, -1.9512743530273435, -9.787496386718749]","[-0.00799387465648855, 0.004529862305343512, -0.002398162396946565]","[68.7, 0.75, -94.05]",Error,Error,Error,0.05706787109375,0.063446044921875,0.048126220703125,25.900000000000006,Error
-23:47:58,50.7,0.5,13.0,Error,Error,Error,Error,Error,Error,Error,"[0.5865794067382812, -2.001552587890625, -9.859322436523437]","[-0.0061286372366412215, 0.00959264958778626, 0.002931087374045802]","[69.0, 1.5, -94.35]",Error,Error,Error,0.0726318359375,0.08074951171875,0.060791015625,25.900000000000006,Error
-23:48:02,51.25,0.5,13.0,Error,Error,Error,Error,Error,Error,Error,"[0.5698199951171875, -1.9632453613281249, -9.849745629882811]","[-0.00985911207633588, 0.008260337145038169, 0.0021316999083969466]","[68.25, 0.0, -94.8]",Error,Error,Error,0.062255859375,0.0692138671875,0.04559326171875,25.950000000000003,Error
-23:48:06,51.25,2.8,13.0,Error,Error,Error,Error,Error,Error,Error,"[0.5506663818359374, -1.9488801513671874, -9.832986218261718]","[-0.00985911207633588, 0.010125574564885496, 0.0013323124427480919]","[70.2, 1.05, -94.05]",Error,Error,Error,0.04150390625,0.046142578125,0.037994384765625,25.950000000000003,Error
-23:48:13,51.8,Error,13.0,Error,Error,Error,07,Error,Error,Error,"[0.584185205078125, -1.956062756347656, -9.804255798339844]","[-0.007727412167938932, 0.014388974381679392, 0.0013323124427480919]","[0.0, 0.0, 0.0]",Error,Error,Error,0.03631591796875,0.040374755859375,0.027862548828125,26.0,Error
-23:48:20,52.35,8.3,13.0,Error,Error,Error,Error,Error,Error,Error,"[0.5961562133789062, -1.9943699829101562, -9.79946739501953]","[-0.008793262122137404, 0.00799387465648855, -0.000799387465648855]","[68.25, 0.75, -94.05]",Error,Error,Error,0.10894775390625,0.121124267578125,0.09625244140625,26.049999999999997,Error
