--- Communications/CSV/16072024/data_235349.csv
@@ -1,11 +1,0 @@
-Date,CPUTemp,CPU Usage,RAM Usage,Latitude,Longitude,Altitude,Satellites,Elevation,Azimuth,UTC Time,Acceleration,Gyro,Magnetic,Pressure,BMP Temp,BMP Altitude,UVA,UVB,UVC,UV Temp,Temperature
-23:53:51,54.55,Error,13.1,40.435938166666666,-3.8834870000000006,Error,Error,Error,Error,215350.00,"[0.5602431884765625, -1.9776105712890624, -9.868899243164062]","[-0.005862174748091604, 0.0061286372366412215, 0.002398162396946565]","[67.64999999999999, 2.1, -94.2]",Error,Error,Error,0.04669189453125,0.046142578125,0.037994384765625,26.849999999999994,Error
-23:53:55,54.0,10.6,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.22266075439453123, -2.3774422485351563, -9.981426721191406]","[0.09539357090076336, 0.06288514729770993, 0.04743032296183207]","[68.7, 2.6999999999999997, -94.8]",Error,Error,Error,0.08819580078125,0.098052978515625,0.068389892578125,26.849999999999994,Error
-23:53:59,55.65,0.3,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.143652099609375, -1.9536685546875, -10.421959826660157]","[-0.06821439706870229, 0.04476569807633588, 0.10818377035114506]","[33.0, 26.7, -49.199999999999996]",Error,Error,Error,0.020751953125,0.0230712890625,0.017730712890625,26.849999999999994,Error
-23:54:03,56.2,0.2,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.316034619140625, -1.8100164550781248, -9.907206469726562]","[-0.01279019945038168, 0.002398162396946565, -0.0013323124427480919]","[32.25, 25.95, -51.0]",Error,Error,Error,0.0726318359375,0.08074951171875,0.058258056640625,26.900000000000006,Error
-23:54:07,54.55,4.5,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.4644751220703125, -1.886630908203125, -9.897629663085937]","[0.008793262122137404, 0.002931087374045802, 0.010125574564885496]","[29.849999999999998, 28.65, -50.25]",Error,Error,Error,0.09857177734375,0.109588623046875,0.0911865234375,26.849999999999994,Error
-23:54:11,54.55,0.3,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.373495458984375, -1.9033903198242186, -9.85692823486328]","[-0.00479632479389313, 0.011990811984732825, -0.0002664624885496183]","[29.7, 28.05, -50.1]",Error,Error,Error,0.04150390625,0.046142578125,0.0303955078125,26.900000000000006,Error
-23:54:15,54.0,0.5,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.19871873779296872, -2.142810485839844, -9.9215716796875]","[0.04583154803053435, -0.02558039890076336, 0.02078407410687023]","[31.2, 26.55, -49.35]",Error,Error,Error,0.103759765625,0.1153564453125,0.0911865234375,26.900000000000006,Error
-23:54:18,55.1,0.5,13.1,Error,Error,Error,12,Error,Error,Error,"[0.40222587890625, -1.8818425048828125, -9.847351428222655]","[-0.01119142451908397, 0.003730474839694657, -0.0005329249770992366]","[30.15, 27.75, -49.8]",Error,Error,Error,0.062255859375,0.074981689453125,0.04559326171875,26.900000000000006,Error
-23:54:22,55.1,0.5,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.47884033203125, -1.89620771484375, -9.830592016601562]","[0.000799387465648855, 0.017586524244274808, -0.007194487190839696]","[28.799999999999997, 26.7, -52.5]",Error,Error,Error,0.062255859375,0.0692138671875,0.0506591796875,26.950000000000003,Error
-23:54:26,55.1,0.3,13.1,Error,Error,Error,Error,Error,Error,Error,"[0.4429273071289062, -1.8842367065429686, -9.883264453125]","[-0.014388974381679392, 0.007727412167938932, 0.000799387465648855]","[29.7, 28.05, -50.85]",Error,Error,Error,0.103759765625,0.1153564453125,0.09625244140625,26.900000000000006,Error
