--- Communications/CSV/16072024/data_235556.csv
@@ -1,10 +1,0 @@
-Date,CPUTemp,CPU Usage,RAM Usage,Latitude,Longitude,Altitude,Satellites,Elevation,Azimuth,UTC Time,Acceleration,Gyro,Magnetic,Pressure,BMP Temp,BMP Altitude,UVA,UVB,UVC,UV Temp,Temperature
-23:56:02,55.65,0.3,12.9,40.43580333333333,-3.8830273333333336,Error,Error,Error,Error,215557.00,"[0.44532150878906246, -1.9033903198242186, -9.849745629882811]","[-0.005595712259541985, 0.008526799633587786, 0.0034640123511450386]","[0.0, 0.0, 0.0]",Error,Error,Error,0.05706787109375,0.063446044921875,0.037994384765625,27.099999999999994,Error
-23:56:09,56.2,14.6,12.9,Error,Error,Error,12,Error,Error,Error,"[0.488417138671875, -1.9033903198242186, -9.801861596679688]","[0.00639509972519084, 0.010924962030534353, 0.0002664624885496183]","[30.75, 27.15, -51.449999999999996]",Error,Error,Error,0.10894775390625,0.121124267578125,0.09625244140625,27.099999999999994,Error
-23:56:12,55.65,0.2,12.9,Error,Error,Error,09,Error,Error,Error,"[0.4836287353515625, -1.898601916503906, -9.866505041503906]","[-0.005062787282442748, 0.007194487190839696, 0.0018652374198473284]","[29.099999999999998, 29.099999999999998, -51.75]",Error,Error,Error,0.04669189453125,0.051910400390625,0.037994384765625,27.150000000000006,Error
-23:56:16,56.2,Error,12.9,Error,Error,Error,Error,Error,Error,Error,"[0.469263525390625, -1.879448303222656, -9.868899243164062]","[-0.004529862305343512, 0.007194487190839696, 0.0018652374198473284]","[31.049999999999997, 27.45, -49.949999999999996]",Error,Error,Error,0.07781982421875,0.086517333984375,0.058258056640625,27.099999999999994,Error
-23:56:20,56.75,14.6,12.9,Error,Error,Error,Error,Error,Error,Error,"[0.4165910888671875, -1.9033903198242186, -9.878476049804688]","[-0.004529862305343512, 0.00959264958778626, -0.0034640123511450386]","[31.049999999999997, 28.349999999999998, -50.25]",Error,Error,Error,0.07781982421875,0.09228515625,0.073455810546875,27.099999999999994,Error
-23:56:24,56.75,1.3,12.9,Error,Error,Error,Error,Error,Error,Error,"[0.4836287353515625, -1.8531120849609375, -9.87608184814453]","[-0.008260337145038169, 0.010924962030534353, -0.00159877493129771]","[30.15, 27.3, -50.1]",Error,Error,Error,0.07781982421875,0.086517333984375,0.060791015625,27.150000000000006,Error
-23:56:28,55.65,0.2,12.9,Error,Error,Error,Error,Error,Error,Error,"[0.4070142822265625, -1.8555062866210936, -9.866505041503906]","[-0.00159877493129771, 0.011990811984732825, 0.002931087374045802]","[29.4, 28.799999999999997, -51.15]",Error,Error,Error,0.1141357421875,0.132659912109375,0.10638427734375,27.200000000000003,Error
-23:56:32,56.2,13.5,13.0,Error,Error,Error,Error,Error,Error,Error,"[0.44532150878906246, -1.886630908203125, -9.85692823486328]","[-0.003996937328244275, 0.008793262122137404, 0.000799387465648855]","[30.75, 26.7, -51.0]",Error,Error,Error,0.0726318359375,0.08074951171875,0.058258056640625,27.099999999999994,Error
-23:56:39,56.2,0.3,13.0,Error,Error,Error,Error,Error,Error,Error,"[0.44053310546874996, -1.89620771484375, -9.854534033203125]","[-0.007727412167938932, 0.0005329249770992366, -0.0010658499541984733]","[0.0, 0.0, 0.0]",Error,Error,Error,0.0726318359375,0.08074951171875,0.058258056640625,27.150000000000006,Error
