--- Communications/CSV/16072024/data_235944.csv
@@ -1,2 +1,0 @@
-Date,CPUTemp,CPU Usage,RAM Usage,Latitude,Longitude,Altitude,Satellites,Elevation,Azimuth,UTC Time,Acceleration,Gyro,Magnetic,Pressure,BMP Temp,BMP Altitude,UVA,UVB,UVC,UV Temp,Temperature
-23:59:50,57.3,0.5,13.0,40.43579066666667,-3.8827823333333336,Error,Error,Error,Error,215945.00,"[0.40940848388671874, -1.9033903198242186, -9.813832604980469]","[-0.005862174748091604, 0.009059724610687023, -0.0018652374198473284]","[29.849999999999998, 27.45, -49.199999999999996]",Error,Error,Error,0.04150390625,0.046142578125,0.027862548828125,27.400000000000006,Error
