--- Communications/CSV/17072024/data_000850.csv
@@ -1,6 +1,0 @@
-Date,CPUTemp,CPU Usage,RAM Usage,Latitude,Longitude,Altitude,Satellites,Elevation,Azimuth,UTC Time,Acceleration,Gyro,Magnetic,Pressure,BMP Temp,BMP Altitude,UVA,UVB,UVC,UV Temp,Temperature
-00:08:52,58.95,0.2,13.3,40.43585916666667,-3.882984833333333,Error,Error,Error,Error,220851.00,"[0.27293898925781246, -1.714248388671875, -9.655815295410155]","[0.0053292497709923675, -0.018119449221374047, -0.002398162396946565]","[31.65, 25.65, -47.55]",Error,Error,Error,0.04669189453125,0.051910400390625,0.04052734375,27.650000000000006,Error
-00:08:59,58.95,5.0,13.3,Error,Error,Error,Error,Error,Error,Error,"[0.13407529296874998, -1.7573440185546874, -9.868899243164062]","[-0.007727412167938932, 0.010924962030534353, 0.0026646248854961837]","[33.0, 26.25, -46.949999999999996]",Error,Error,Error,0.12451171875,0.144195556640625,0.1114501953125,27.650000000000006,Error
-00:09:06,58.4,0.2,13.3,Error,Error,Error,Error,Error,Error,Error,"[0.11731588134765625, -1.7621324218749999, -9.854534033203125]","[-0.008526799633587786, 0.007194487190839696, 0.002398162396946565]","[31.95, 27.75, -49.35]",Error,Error,Error,0.10894775390625,0.12689208984375,0.09625244140625,27.650000000000006,Error
-00:09:14,59.5,0.2,13.3,Error,Error,Error,Error,Error,Error,Error,"[0.1484405029296875, -1.7310078002929687, -9.871293444824218]","[-0.011724349496183208, 0.0005329249770992366, 0.00159877493129771]","[30.9, 25.65, -47.1]",Error,Error,Error,0.00518798828125,0.005767822265625,0.007598876953125,27.650000000000006,Error
-00:09:17,58.95,0.2,13.3,Error,Error,Error,Error,Error,Error,Error,"[0.09816226806640624, -1.695094775390625, -10.029310754394531]","[-0.0002664624885496183, 0.007460949679389314, -0.0013323124427480919]","[31.5, 25.2, -48.0]",Error,Error,Error,0.0830078125,0.09228515625,0.06585693359375,27.650000000000006,Error
