--- Communications/CSV/17072024/data_000947.csv
@@ -1,3 +1,0 @@
-Date,CPUTemp,CPU Usage,RAM Usage,Latitude,Longitude,Altitude,Satellites,Elevation,Azimuth,UTC Time,Acceleration,Gyro,Magnetic,Pressure,BMP Temp,BMP Altitude,UVA,UVB,UVC,UV Temp,Temperature
-00:09:53,58.95,0.2,13.3,40.435872,-3.882985333333334,Error,Error,Error,Error,220948.00,"[0.086191259765625, -1.7621324218749999, -9.888052856445311]","[-0.008526799633587786, 0.0034640123511450386, 0.0013323124427480919]","[0.0, 0.0, 0.0]",Error,Error,Error,0.02593994140625,0.028839111328125,0.022796630859375,27.650000000000006,Error
-00:10:00,58.95,4.5,13.4,Error,Error,Error,12,Error,Error,Error,"[0.11731588134765625, -1.7286135986328124, -9.885658654785155]","[-0.010125574564885496, 0.00479632479389313, 0.0]","[32.699999999999996, 24.9, -46.35]",Error,Error,Error,0.09857177734375,0.109588623046875,0.0810546875,27.650000000000006,Error
