--- Communications/CSV/17072024/data_002800.csv
@@ -1,12 +1,0 @@
-Date,CPUTemp,CPU Usage,RAM Usage,Latitude,Longitude,Altitude,Satellites,Elevation,Azimuth,UTC Time,Acceleration,Gyro,Magnetic,Pressure,BMP Temp,BMP Altitude,UVA,UVB,UVC,UV Temp,Temperature
-00:28:02,58.95,5.5,13.3,40.435671166666665,-3.882790166666666,Error,Error,Error,Error,222801.00,"[0.19632453613281248, -1.7022773803710936, -9.730035546875]","[0.00799387465648855, 0.019984686641221373, -0.0034640123511450386]","[33.6, 25.5, -48.449999999999996]",Error,Error,Error,0.07781982421875,0.086517333984375,0.0709228515625,27.849999999999994,Error
-00:28:06,59.5,0.2,13.3,Error,Error,Error,Error,Error,Error,Error,"[0.15562310791015624, -1.7908628417968748, -9.885658654785155]","[-0.007194487190839696, 0.003996937328244275, -0.0005329249770992366]","[32.25, 26.25, -48.0]",Error,Error,Error,0.0830078125,0.09228515625,0.063323974609375,27.799999999999997,Error
-00:28:10,58.4,0.3,13.3,Error,Error,Error,Error,Error,Error,Error,"[0.15083470458984374, -1.7884686401367187, -9.880870251464843]","[-0.005062787282442748, 0.005062787282442748, -0.002931087374045802]","[31.95, 26.55, -47.85]",Error,Error,Error,0.062255859375,0.0692138671875,0.053192138671875,27.849999999999994,Error
-00:28:14,58.4,12.1,13.3,Error,Error,Error,Error,Error,Error,Error,"[0.12449848632812499, -1.7094599853515624, -9.832986218261718]","[-0.007460949679389314, 0.0034640123511450386, -0.0013323124427480919]","[33.15, 27.3, -49.05]",Error,Error,Error,0.08819580078125,0.098052978515625,0.063323974609375,27.849999999999994,Error
-00:28:18,58.4,Error,13.3,Error,Error,Error,Error,Error,Error,Error,"[0.13168109130859373, -1.7549498168945312, -9.95987890625]","[-0.012257274473282443, 0.006928024702290077, -0.0013323124427480919]","[30.9, 26.849999999999998, -47.55]",Error,Error,Error,0.05706787109375,0.063446044921875,0.053192138671875,27.900000000000006,Error
-00:28:22,58.4,0.5,13.3,Error,Error,Error,Error,Error,Error,Error,"[0.2442085693359375, -1.714248388671875, -9.931148486328125]","[-0.018652374198473286, 0.014388974381679392, -0.003996937328244275]","[32.25, 24.3, -49.199999999999996]",Error,Error,Error,0.05706787109375,0.063446044921875,0.04559326171875,27.849999999999994,Error
-00:28:26,59.5,14.6,13.3,Error,Error,Error,Error,Error,Error,Error,"[0.7829039428710937, -1.5538368774414062, -9.313444458007812]","[-0.002398162396946565, 0.000799387465648855, -0.003730474839694657]","[35.25, 18.0, -50.25]",Error,Error,Error,0.0726318359375,0.08074951171875,0.05572509765625,27.799999999999997,Error
-00:28:29,58.95,0.2,13.3,Error,Error,Error,09,Error,Error,Error,"[0.35673604736328124, -1.9847931762695312, -9.811438403320311]","[0.014122511893129772, 0.008793262122137404, 0.00959264958778626]","[34.8, 18.0, -48.75]",Error,Error,Error,0.062255859375,0.0692138671875,0.04559326171875,27.900000000000006,Error
-00:28:33,58.95,0.5,13.3,Error,Error,Error,Error,Error,Error,Error,"[0.5602431884765625, -1.9225439331054686, -9.897629663085937]","[0.0005329249770992366, 0.006928024702290077, 0.00159877493129771]","[35.1, 19.05, -49.8]",Error,Error,Error,0.02593994140625,0.028839111328125,0.022796630859375,27.849999999999994,Error
-00:28:40,59.5,0.2,13.3,Error,Error,Error,Error,Error,Error,Error,"[0.5793968017578125, -1.9105729248046874, -9.852139831542969]","[-0.002931087374045802, 0.002398162396946565, 0.003996937328244275]","[0.0, 0.0, 0.0]",Error,Error,Error,0.062255859375,0.0692138671875,0.03546142578125,27.799999999999997,Error
-00:28:47,58.95,0.2,13.3,Error,Error,Error,Error,Error,Error,Error,"[0.6105214233398437, -1.7860744384765623, -9.907206469726562]","[-0.005062787282442748, 0.02398162396946565, -0.00159877493129771]","[35.85, 19.349999999999998, -48.6]",Error,Error,Error,0.0311279296875,0.03460693359375,0.0303955078125,27.900000000000006,Error
