--- Communications/CSV/17072024/data_132547.csv
@@ -1,16 +1,0 @@
-Date,CPUTemp,CPU Usage,RAM Usage,Latitude,Longitude,Altitude,Satellites,Elevation,Azimuth,UTC Time,Acceleration,Gyro,Magnetic,Pressure,BMP Temp,BMP Altitude,UVA,UVB,UVC,UV Temp,Temperature
-13:25:49,80.95,Error,13.2,40.4355505,-3.8824009999999998,Error,Error,Error,Error,112548.00,"[0.9911994873046874, -0.7158662963867187, -10.01015714111328]","[-0.009059724610687023, 0.0061286372366412215, 0.000799387465648855]","[-32.699999999999996, 26.849999999999998, -40.5]",Error,Error,Error,2.87933349609375,0.19610595703125,0.12664794921875,29.900000000000006,Error
-13:25:53,82.05,14.6,13.2,Error,Error,Error,Error,Error,Error,Error,"[0.9744400756835937, -0.699106884765625, -10.002974536132811]","[-0.005862174748091604, 0.008793262122137404, 0.004529862305343512]","[-32.25, 27.9, -41.1]",Error,Error,Error,2.94677734375,0.271087646484375,0.1925048828125,30.0,Error
-13:25:57,83.15,Error,13.2,Error,Error,Error,Error,Error,Error,Error,"[1.0366893188476562, -0.6560112548828124, -9.950302099609374]","[-0.005062787282442748, 0.00959264958778626, 0.0013323124427480919]","[-31.2, 26.849999999999998, -39.3]",Error,Error,Error,2.90008544921875,0.21917724609375,0.149444580078125,29.950000000000003,Error
-13:26:01,80.4,1.0,13.2,Error,Error,Error,Error,Error,Error,Error,"[0.9145850341796874, -0.6943184814453125, -9.926360083007811]","[-0.00799387465648855, 0.007727412167938932, 0.002398162396946565]","[-30.75, 27.15, -41.1]",Error,Error,Error,2.84820556640625,0.14996337890625,0.093719482421875,30.0,Error
-13:26:05,79.85,13.4,13.2,Error,Error,Error,Error,Error,Error,Error,"[1.005564697265625, -0.6272808349609375, -9.947907897949218]","[-0.004263399816793893, 0.010658499541984735, -0.003730474839694657]","[-30.75, 26.25, -40.35]",Error,Error,Error,2.86895751953125,0.167266845703125,0.103851318359375,30.0,Error
-13:26:09,82.6,0.2,13.2,Error,Error,Error,Error,Error,Error,Error,"[0.9864110839843749, -0.622492431640625, -9.991003527832031]","[-0.008260337145038169, 0.010392037053435116, -0.002931087374045802]","[-31.5, 25.5, -40.65]",Error,Error,Error,2.88970947265625,0.1845703125,0.124114990234375,29.950000000000003,Error
-13:26:13,79.85,0.2,13.2,Error,Error,Error,Error,Error,Error,Error,"[0.9744400756835937, -0.6560112548828124, -9.981426721191406]","[-0.005862174748091604, 0.005595712259541985, -0.0013323124427480919]","[-29.549999999999997, 26.7, -39.9]",Error,Error,Error,2.843017578125,0.1153564453125,0.07598876953125,29.950000000000003,Error
-13:26:16,79.3,6.5,13.2,Error,Error,Error,08,Error,Error,Error,"[1.0510545288085937, -0.6919242797851562, -9.916783276367187]","[-0.00479632479389313, 0.00959264958778626, 0.002931087374045802]","[-32.25, 26.7, -41.1]",Error,Error,Error,2.84820556640625,0.132659912109375,0.073455810546875,29.900000000000006,Error
-13:26:20,81.5,0.3,13.2,Error,Error,Error,Error,Error,Error,Error,"[1.0749965454101562, -0.6512228515624999, -9.969455712890625]","[-0.010392037053435116, 0.011990811984732825, -0.004263399816793893]","[-30.45, 26.849999999999998, -39.0]",Error,Error,Error,2.8948974609375,0.236480712890625,0.154510498046875,29.950000000000003,Error
-13:26:27,79.3,Error,13.2,Error,Error,Error,Error,Error,Error,Error,"[1.0127473022460938, -0.6775590698242188, -9.969455712890625]","[-0.008793262122137404, 0.013856049404580155, 0.0002664624885496183]","[0.0, 0.0, 0.0]",Error,Error,Error,2.82745361328125,0.155731201171875,0.101318359375,30.0,Error
-13:26:34,81.5,1.3,13.2,Error,Error,Error,Error,Error,Error,Error,"[1.0031704956054688, -0.6584054565429687, -9.91438907470703]","[-0.00639509972519084, 0.014655436870229008, -0.0021316999083969466]","[-30.9, 28.5, -38.25]",Error,Error,Error,2.85858154296875,0.167266845703125,0.103851318359375,29.950000000000003,Error
-13:26:38,78.75,0.5,13.2,Error,Error,Error,Error,Error,Error,Error,"[0.9121908325195311, -0.6727706665039062, -9.962273107910155]","[-0.00639509972519084, 0.008260337145038169, 0.002398162396946565]","[-29.7, 28.5, -41.25]",Error,Error,Error,2.87933349609375,0.190338134765625,0.12158203125,30.0,Error
-13:26:42,80.4,11.3,13.2,Error,Error,Error,Error,Error,Error,Error,"[0.9504980590820312, -0.6416460449218749, -10.01015714111328]","[-0.007460949679389314, 0.005862174748091604, -0.0005329249770992366]","[-31.5, 27.9, -39.9]",Error,Error,Error,2.87933349609375,0.213409423828125,0.144378662109375,30.0,Error
-13:26:46,82.05,0.3,13.2,Error,Error,Error,Error,Error,Error,Error,"[1.0319009155273438, -0.63206923828125, -9.97184991455078]","[-0.005862174748091604, 0.007727412167938932, 0.0002664624885496183]","[-30.299999999999997, 27.9, -39.6]",Error,Error,Error,2.8948974609375,0.230712890625,0.159576416015625,29.950000000000003,Error
-13:26:50,79.3,0.2,13.2,Error,Error,Error,Error,Error,Error,Error,"[0.9720458740234375, -0.6560112548828124, -9.955090502929687]","[-0.007460949679389314, 0.00479632479389313, -0.0018652374198473284]","[-30.15, 27.75, -39.75]",Error,Error,Error,2.884521484375,0.190338134765625,0.10638427734375,30.0,Error
