--- Communications/CSV/17072024/data_134820.csv
@@ -1,4 +1,0 @@
-Date,CPUTemp,CPU Usage,RAM Usage,Latitude,Longitude,Altitude,Satellites,Elevation,Azimuth,UTC Time,Acceleration,Gyro,Magnetic,Pressure,BMP Temp,BMP Altitude,UVA,UVB,UVC,UV Temp,Temperature
-13:48:22,79.3,Error,13.3,40.4356585,-3.8826038333333335,Error,Error,Error,Error,114821.00,"[0.9624690673828125, -0.5506663818359374, -9.991003527832031]","[-0.00479632479389313, 0.00319754986259542, -0.003996937328244275]","[-30.45, 21.75, -39.75]",Error,Error,Error,2.94677734375,0.248016357421875,0.17730712890625,30.150000000000006,Error
-13:48:26,79.3,0.3,13.3,Error,Error,Error,13,Error,Error,Error,"[0.93852705078125, -0.612915625, -9.983820922851562]","[-0.0061286372366412215, 0.008260337145038169, 0.002398162396946565]","[-28.65, 21.75, -40.05]",Error,Error,Error,2.884521484375,0.14996337890625,0.08612060546875,30.099999999999994,Error
-13:48:30,77.65,Error,13.3,Error,Error,Error,13,Error,Error,Error,"[0.9840168823242187, -0.5746083984375, -9.964667309570311]","[-0.005062787282442748, 0.002398162396946565, -0.0021316999083969466]","[-30.299999999999997, 20.4, -41.4]",Error,Error,Error,2.967529296875,0.224945068359375,0.14691162109375,30.099999999999994,Error
