--- Communications/CSV/17072024/data_134834.csv
@@ -1,5 +1,0 @@
-Date,CPUTemp,CPU Usage,RAM Usage,Latitude,Longitude,Altitude,Satellites,Elevation,Azimuth,UTC Time,Acceleration,Gyro,Magnetic,Pressure,BMP Temp,BMP Altitude,UVA,UVB,UVC,UV Temp,Temperature
-13:48:36,80.95,0.2,13.3,40.435673666666666,-3.8826863333333335,Error,Error,Error,Error,114835.00,"[0.9576806640625, -0.5961562133789062, -9.950302099609374]","[-0.007727412167938932, 0.0013323124427480919, 0.004263399816793893]","[-31.65, 20.25, -41.25]",Error,Error,Error,2.97271728515625,0.24224853515625,0.162109375,30.099999999999994,Error
-13:48:40,77.65,0.2,13.3,Error,Error,Error,13,Error,Error,Error,"[0.97683427734375, -0.603338818359375, -9.986215124511718]","[-0.009326187099236643, 0.00985911207633588, 0.0026646248854961837]","[-30.75, 20.849999999999998, -39.15]",Error,Error,Error,2.94158935546875,0.248016357421875,0.16717529296875,30.099999999999994,Error
-13:48:44,77.65,9.3,13.3,Error,Error,Error,13,Error,Error,Error,"[0.9193734375, -0.5746083984375, -9.995791931152343]","[-0.00959264958778626, 0.003730474839694657, -0.002931087374045802]","[-30.75, 20.849999999999998, -40.35]",Error,Error,Error,2.87933349609375,0.1614990234375,0.093719482421875,30.049999999999997,Error
-13:48:48,80.4,Error,13.3,Error,Error,Error,13,Error,Error,Error,"[0.9720458740234375, -0.622492431640625, -9.988609326171874]","[-0.006928024702290077, 0.006661562213740459, -0.0021316999083969466]","[-30.9, 21.75, -41.25]",Error,Error,Error,2.8741455078125,0.14996337890625,0.078521728515625,30.099999999999994,Error
