--- Communications/CSV/17072024/data_140308.csv
@@ -1,17 +1,0 @@
-Date,CPUTemp,CPU Usage,RAM Usage,Latitude,Longitude,Altitude,Satellites,Elevation,Azimuth,UTC Time,Acceleration,Gyro,Magnetic,Pressure,BMP Temp,BMP Altitude,UVA,UVB,UVC,UV Temp,Temperature
-14:03:10,79.85,8.0,13.3,40.43573466666667,-3.8831058333333335,Error,Error,Error,Error,120309.00,"[1.0678139404296874, -0.7110778930664062, -9.957484704589843]","[-0.007460949679389314, 0.0034640123511450386, -0.0013323124427480919]","[-31.2, 33.6, -34.949999999999996]",Error,Error,Error,3.36181640625,0.201873779296875,0.12158203125,30.200000000000003,Error
-14:03:14,79.85,1.0,13.3,Error,Error,Error,14,Error,Error,Error,"[1.01514150390625, -0.6512228515624999, -9.962273107910155]","[-0.008526799633587786, 0.006661562213740459, -0.0026646248854961837]","[-31.049999999999997, 34.199999999999996, -34.05]",Error,Error,Error,3.40850830078125,0.224945068359375,0.149444580078125,30.200000000000003,Error
-14:03:18,78.2,0.2,13.3,Error,Error,Error,14,Error,Error,Error,"[1.0175357055664063, -0.679953271484375, -9.943119494628906]","[-0.010658499541984735, 0.005595712259541985, 0.008793262122137404]","[-31.5, 33.75, -35.55]",Error,Error,Error,3.3306884765625,0.144195556640625,0.0810546875,30.25,Error
-14:03:22,77.65,1.5,13.3,Error,Error,Error,14,Error,Error,Error,"[0.9840168823242187, -0.6919242797851562, -9.955090502929687]","[-0.007460949679389314, 0.00639509972519084, 0.000799387465648855]","[-31.799999999999997, 32.55, -33.6]",Error,Error,Error,3.341064453125,0.190338134765625,0.10638427734375,30.25,Error
-14:03:26,79.3,Error,13.3,Error,Error,Error,14,Error,Error,Error,"[0.9864110839843749, -0.6847416748046875, -9.97903251953125]","[-0.0034640123511450386, 0.00159877493129771, 0.0]","[-30.45, 33.15, -34.199999999999996]",Error,Error,Error,3.424072265625,0.17303466796875,0.11651611328125,30.25,Error
-14:03:30,78.75,1.2,13.3,Error,Error,Error,14,Error,Error,Error,"[1.0366893188476562, -0.6488286499023437, -9.952696301269532]","[-0.007727412167938932, 0.00799387465648855, -0.00479632479389313]","[-32.4, 33.15, -34.949999999999996]",Error,Error,Error,3.34625244140625,0.132659912109375,0.0810546875,30.299999999999997,Error
-14:03:34,78.2,Error,13.3,Error,Error,Error,Error,Error,Error,Error,"[0.9983820922851562, -0.7038952880859375, -9.89523546142578]","[-0.0061286372366412215, 0.002931087374045802, 0.0034640123511450386]","[-31.5, 33.449999999999996, -33.3]",Error,Error,Error,3.33587646484375,0.12689208984375,0.0709228515625,30.299999999999997,Error
-14:03:37,79.3,Error,13.3,Error,Error,Error,09,Error,Error,Error,"[1.0079588989257813, -0.6727706665039062, -9.931148486328125]","[-0.007460949679389314, 0.011724349496183208, 0.0010658499541984733]","[-30.9, 33.6, -33.75]",Error,Error,Error,3.36700439453125,0.155731201171875,0.098785400390625,30.200000000000003,Error
-14:03:41,79.3,0.3,13.3,Error,Error,Error,14,Error,Error,Error,"[0.9888052856445312, -0.70868369140625, -9.969455712890625]","[-0.005862174748091604, 0.00479632479389313, -0.0034640123511450386]","[-30.299999999999997, 34.199999999999996, -35.55]",Error,Error,Error,3.36700439453125,0.1614990234375,0.101318359375,30.299999999999997,Error
-14:03:48,79.3,6.3,13.3,Error,Error,Error,14,Error,Error,Error,"[0.9888052856445312, -0.7038952880859375, -9.926360083007811]","[-0.00799387465648855, 0.00639509972519084, -0.0013323124427480919]","[-30.75, 33.75, -34.35]",Error,Error,Error,3.382568359375,0.1845703125,0.1317138671875,30.25,Error
-14:03:55,79.3,0.3,13.3,Error,Error,Error,14,Error,Error,Error,"[1.0606313354492187, -0.6847416748046875, -9.911994873046874]","[-0.008260337145038169, 0.007460949679389314, -0.000799387465648855]","[-31.95, 33.15, -34.949999999999996]",Error,Error,Error,3.4136962890625,0.17303466796875,0.108917236328125,30.299999999999997,Error
-14:03:59,78.75,14.4,13.3,Error,Error,Error,14,Error,Error,Error,"[1.0438719238281249, -0.7062894897460937, -9.95987890625]","[-0.0130566619389313, 0.0005329249770992366, 0.0010658499541984733]","[-31.2, 33.15, -34.5]",Error,Error,Error,3.37738037109375,0.190338134765625,0.11651611328125,30.25,Error
-14:04:03,80.4,1.2,13.3,Error,Error,Error,14,Error,Error,Error,"[1.0606313354492187, -0.6847416748046875, -9.957484704589843]","[-0.008260337145038169, 0.00479632479389313, 0.002398162396946565]","[-31.799999999999997, 33.449999999999996, -35.1]",Error,Error,Error,3.36181640625,0.167266845703125,0.101318359375,30.299999999999997,Error
-14:04:07,78.75,0.2,13.3,Error,Error,Error,14,Error,Error,Error,"[0.9792284790039062, -0.7038952880859375, -9.94072529296875]","[-0.01279019945038168, 0.002398162396946565, 0.0034640123511450386]","[-31.2, 33.15, -33.449999999999996]",Error,Error,Error,3.4136962890625,0.2076416015625,0.1317138671875,30.25,Error
-14:04:11,77.1,10.6,13.3,Error,Error,Error,14,Error,Error,Error,"[1.0223241088867188, -0.718260498046875, -9.969455712890625]","[-0.011457887007633588, 0.0053292497709923675, -0.0034640123511450386]","[-29.849999999999998, 34.199999999999996, -34.05]",Error,Error,Error,3.38775634765625,0.201873779296875,0.124114990234375,30.299999999999997,Error
-14:04:15,80.95,Error,13.3,Error,Error,Error,14,Error,Error,Error,"[0.948103857421875, -0.7302315063476562, -9.955090502929687]","[-0.00639509972519084, 0.003996937328244275, 0.002931087374045802]","[-32.25, 31.799999999999997, -34.35]",Error,Error,Error,3.34625244140625,0.144195556640625,0.083587646484375,30.299999999999997,Error
