--- Communications/CSV/17072024/data_145704.csv
@@ -1,9 +1,0 @@
-Date,CPUTemp,CPU Usage,RAM Usage,Latitude,Longitude,Altitude,Satellites,Elevation,Azimuth,UTC Time,Acceleration,Gyro,Magnetic,UVA,UVB,UVC,UV Temp,Temperature
-14:57:10,81.5,1.2,13.4,Error,Error,Error,Error,Error,Error,Error,"[0.9600748657226562, -0.7158662963867187, -9.945513696289062]","[-0.004263399816793893, 0.005062787282442748, -0.000799387465648855]","[-30.45, 34.05, -34.949999999999996]",3.91693115234375,0.14996337890625,0.083587646484375,30.549999999999997,Error
-14:57:16,77.65,0.2,13.4,Error,Error,Error,Error,Error,Error,Error,"[0.97683427734375, -0.7110778930664062, -9.94072529296875]","[-0.005062787282442748, 0.007460949679389314, 0.0005329249770992366]","[-30.299999999999997, 33.0, -34.35]",4.02587890625,0.236480712890625,0.159576416015625,30.599999999999994,Error
-14:57:23,79.85,0.5,13.4,Error,Error,Error,Error,Error,Error,Error,"[0.93852705078125, -0.699106884765625, -9.974244116210937]","[-0.012257274473282443, 0.00799387465648855, -0.000799387465648855]","[-31.799999999999997, 33.449999999999996, -34.05]",3.95843505859375,0.21917724609375,0.134246826171875,30.549999999999997,Error
-14:57:30,78.2,0.3,13.4,Error,Error,Error,Error,Error,Error,Error,"[1.0223241088867188, -0.7230489013671875, -9.995791931152343]","[-0.008526799633587786, 0.004263399816793893, -0.002931087374045802]","[-31.799999999999997, 33.449999999999996, -35.1]",4.1192626953125,0.305694580078125,0.20263671875,30.549999999999997,Error
-14:57:37,79.85,0.5,13.4,Error,Error,Error,Error,Error,Error,Error,"[1.0031704956054688, -0.6751648681640625, -9.926360083007811]","[-0.003730474839694657, 0.008260337145038169, 0.002398162396946565]","[-31.65, 33.15, -33.0]",3.97918701171875,0.2076416015625,0.12664794921875,30.599999999999994,Error
-14:57:44,77.1,14.6,13.4,Error,Error,Error,Error,Error,Error,Error,"[0.9457096557617187, -0.7278373046875, -9.94072529296875]","[-0.01119142451908397, 0.003730474839694657, -0.000799387465648855]","[-31.95, 33.0, -33.3]",3.96881103515625,0.24224853515625,0.13677978515625,30.549999999999997,Error
-14:57:54,78.2,0.2,13.5,Error,Error,Error,Error,Error,Error,Error,"[0.9648632690429687, -0.7254431030273437, -9.943119494628906]","[-0.007727412167938932, 0.005595712259541985, 0.0]","[-31.95, 31.65, -36.0]",3.92730712890625,0.201873779296875,0.12664794921875,30.549999999999997,Error
-14:58:01,80.4,0.3,13.5,Error,Error,Error,Error,Error,Error,Error,"[0.9864110839843749, -0.7398083129882812, -9.97184991455078]","[-0.0061286372366412215, 0.004529862305343512, -0.0026646248854961837]","[-30.9, 34.05, -33.75]",4.005126953125,0.248016357421875,0.159576416015625,30.599999999999994,Error
