--- Communications/CSV/17072024/data_150542.csv
@@ -1,10 +1,0 @@
-Date,CPUTemp,CPU Usage,RAM Usage,GPS,IMU,UV,Dallas Temperature
-2024-07-17 15:05:47,{'Temperature': 77.1},{'Usage': 0.0},{'Usage': 13.4},Error,"{'Acceleration': [1.024718310546875, -0.7876923461914062, -9.904812268066406], 'Gyro': [-0.009059724610687023, 0.006928024702290077, 0.00479632479389313], 'Magnetic': [-33.0, 31.5, -36.15]}","{'UVA': 3.79241943359375, 'UVB': 0.19610595703125, 'UVC': 0.129180908203125, 'Temperature': 30.349999999999994}",Error
-2024-07-17 15:05:54,{'Temperature': 75.45},{'Usage': 0.3},{'Usage': 13.5},Error,"{'Acceleration': [1.0462661254882812, -0.7398083129882812, -9.89523546142578], 'Gyro': [-0.0130566619389313, 0.002931087374045802, 0.0053292497709923675], 'Magnetic': [-33.75, 31.799999999999997, -36.9]}","{'UVA': 3.75091552734375, 'UVB': 0.144195556640625, 'UVC': 0.08612060546875, 'Temperature': 30.349999999999994}",Error
-2024-07-17 15:06:01,{'Temperature': 76.55},{'Usage': 0.0},{'Usage': 13.5},Error,"{'Acceleration': [1.0941501586914062, -0.7206546997070312, -9.9215716796875], 'Gyro': [-0.00639509972519084, 0.00985911207633588, -0.0034640123511450386], 'Magnetic': [-30.15, 30.0, -36.0]}","{'UVA': 3.8079833984375, 'UVB': 0.201873779296875, 'UVC': 0.1317138671875, 'Temperature': 30.299999999999997}",Error
-2024-07-17 15:06:08,{'Temperature': 74.35},{'Usage': 0.3},{'Usage': 13.5},Error,"{'Acceleration': [1.0366893188476562, -0.7302315063476562, -9.938331091308592], 'Gyro': [-0.006928024702290077, 0.005862174748091604, 0.0010658499541984733], 'Magnetic': [-31.5, 31.049999999999997, -36.15]}","{'UVA': 3.80279541015625, 'UVB': 0.271087646484375, 'UVC': 0.179840087890625, 'Temperature': 30.299999999999997}",Error
-2024-07-17 15:06:15,{'Temperature': 77.1},{'Usage': 0.5},{'Usage': 13.5},Error,"{'Acceleration': [1.0295067138671874, -0.7062894897460937, -9.911994873046874], 'Gyro': [-0.010392037053435116, 0.007194487190839696, -0.0005329249770992366], 'Magnetic': [-32.699999999999996, 32.4, -35.25]}","{'UVA': 3.77166748046875, 'UVB': 0.167266845703125, 'UVC': 0.103851318359375, 'Temperature': 30.299999999999997}",Error
-2024-07-17 15:06:21,{'Temperature': 74.35},{'Usage': 9.3},{'Usage': 13.5},Error,"{'Acceleration': [1.0271125122070313, -0.6655880615234374, -9.957484704589843], 'Gyro': [-0.008260337145038169, 0.00159877493129771, 0.004529862305343512], 'Magnetic': [-32.4, 31.95, -36.449999999999996]}","{'UVA': 3.77685546875, 'UVB': 0.259552001953125, 'UVC': 0.169708251953125, 'Temperature': 30.299999999999997}",Error
-2024-07-17 15:06:28,{'Temperature': 75.45},{'Usage': 1.2},{'Usage': 13.5},Error,"{'Acceleration': [1.0486603271484374, -0.7278373046875, -9.923965881347655], 'Gyro': [-0.0018652374198473284, 0.00639509972519084, -0.0010658499541984733], 'Magnetic': [-31.95, 32.85, -36.75]}","{'UVA': 3.756103515625, 'UVB': 0.14996337890625, 'UVC': 0.093719482421875, 'Temperature': 30.299999999999997}",Error
-2024-07-17 15:06:35,{'Temperature': 77.1},{'Usage': 6.8},{'Usage': 13.5},Error,"{'Acceleration': [1.0199299072265624, -0.7110778930664062, -9.95987890625], 'Gyro': [-0.007460949679389314, 0.0061286372366412215, -0.004529862305343512], 'Magnetic': [-30.45, 30.9, -37.199999999999996]}","{'UVA': 3.81317138671875, 'UVB': 0.230712890625, 'UVC': 0.14691162109375, 'Temperature': 30.25}",Error
-2024-07-17 15:06:42,{'Temperature': 74.9},{'Usage': 0.3},{'Usage': 13.5},Error,"{'Acceleration': [1.0271125122070313, -0.7110778930664062, -9.981426721191406], 'Gyro': [-0.007460949679389314, 0.008526799633587786, 0.0034640123511450386], 'Magnetic': [-32.55, 31.5, -35.85]}","{'UVA': 3.8079833984375, 'UVB': 0.2076416015625, 'UVC': 0.139312744140625, 'Temperature': 30.25}",Error
