--- Communications/CSV/17072024/data_175453.csv
@@ -1,20 +1,0 @@
-Date,CPUTemp,CPU Usage,RAM Usage,GPS,IMU,UV,Dallas Temperature
-2024-07-17 17:54:58,{'Temperature': 63.35},{'Usage': 0.0},{'Usage': 12.7},Error,"{'Acceleration': [1.0534487304687499, -0.6775590698242188, -9.95987890625], 'Gyro': [-0.014388974381679392, 0.003996937328244275, -0.004263399816793893], 'Magnetic': [-28.65, 10.799999999999999, -44.25]}","{'UVA': 0.166015625, 'UVB': 0.121124267578125, 'UVC': 0.09625244140625, 'Temperature': 27.75}",Error
-2024-07-17 17:55:05,{'Temperature': 62.25},{'Usage': 0.0},{'Usage': 12.8},Error,"{'Acceleration': [1.0271125122070313, -0.6775590698242188, -9.969455712890625], 'Gyro': [-0.007727412167938932, 0.00159877493129771, 0.00159877493129771], 'Magnetic': [-28.05, 9.45, -43.199999999999996]}","{'UVA': 0.103759765625, 'UVB': 0.046142578125, 'UVC': 0.04052734375, 'Temperature': 27.75}",Error
-2024-07-17 17:55:15,{'Temperature': 60.6},{'Usage': 0.0},{'Usage': 12.8},Error,"{'Acceleration': [1.0271125122070313, -0.6416460449218749, -9.900023864746093], 'Gyro': [-0.011457887007633588, 0.008260337145038169, 0.003730474839694657], 'Magnetic': [-28.95, 12.75, -42.75]}","{'UVA': 0.145263671875, 'UVB': 0.1153564453125, 'UVC': 0.083587646484375, 'Temperature': 27.75}",Error
-2024-07-17 17:55:22,{'Temperature': 61.15},{'Usage': 0.0},{'Usage': 12.8},Error,"{'Acceleration': [1.0845733520507812, -0.6416460449218749, -9.976638317871092], 'Gyro': [-0.003730474839694657, 0.012523736961832063, -0.0026646248854961837], 'Magnetic': [-27.9, 9.6, -43.5]}","{'UVA': 0.12451171875, 'UVB': 0.074981689453125, 'UVC': 0.060791015625, 'Temperature': 27.700000000000003}",Error
-2024-07-17 17:55:29,{'Temperature': 62.25},{'Usage': 0.0},{'Usage': 12.8},Error,"{'Acceleration': [1.0965443603515623, -0.6727706665039062, -9.991003527832031], 'Gyro': [-0.004263399816793893, 0.004529862305343512, -0.000799387465648855], 'Magnetic': [-28.65, 10.799999999999999, -43.5]}","{'UVA': 0.16082763671875, 'UVB': 0.144195556640625, 'UVC': 0.098785400390625, 'Temperature': 27.75}",Error
-2024-07-17 17:55:36,{'Temperature': 60.6},{'Usage': 0.2},{'Usage': 12.8},Error,"{'Acceleration': [1.0558429321289062, -0.6512228515624999, -9.928754284667969], 'Gyro': [-0.009326187099236643, 0.00799387465648855, 0.0005329249770992366], 'Magnetic': [-29.7, 10.35, -44.85]}","{'UVA': 0.1141357421875, 'UVB': 0.063446044921875, 'UVC': 0.053192138671875, 'Temperature': 27.700000000000003}",Error
-2024-07-17 17:55:43,{'Temperature': 61.7},{'Usage': 0.0},{'Usage': 12.8},Error,"{'Acceleration': [0.9959878906249999, -0.6943184814453125, -9.95987890625], 'Gyro': [-0.004529862305343512, 0.00159877493129771, 0.0005329249770992366], 'Magnetic': [-29.4, 9.6, -43.5]}","{'UVA': 0.18157958984375, 'UVB': 0.167266845703125, 'UVC': 0.12664794921875, 'Temperature': 27.75}",Error
-2024-07-17 17:55:49,{'Temperature': 60.05},{'Usage': 0.5},{'Usage': 12.8},Error,"{'Acceleration': [1.0630255371093749, -0.6967126831054687, -9.95987890625], 'Gyro': [-0.008526799633587786, 0.007727412167938932, 0.000799387465648855], 'Magnetic': [-28.799999999999997, 9.75, -43.65]}","{'UVA': 0.18157958984375, 'UVB': 0.138427734375, 'UVC': 0.10638427734375, 'Temperature': 27.700000000000003}",Error
-2024-07-17 17:55:56,{'Temperature': 62.8},{'Usage': 0.0},{'Usage': 12.8},Error,"{'Acceleration': [1.0199299072265624, -0.6943184814453125, -9.983820922851562], 'Gyro': [-0.00985911207633588, 0.00639509972519084, 0.0005329249770992366], 'Magnetic': [-30.15, 9.9, -43.5]}","{'UVA': 0.166015625, 'UVB': 0.121124267578125, 'UVC': 0.098785400390625, 'Temperature': 27.700000000000003}",Error
-2024-07-17 17:56:03,{'Temperature': 59.5},{'Usage': 14.5},{'Usage': 12.8},Error,"{'Acceleration': [0.9792284790039062, -0.6655880615234374, -9.97184991455078], 'Gyro': [-0.0053292497709923675, 0.0002664624885496183, -0.0013323124427480919], 'Magnetic': [-28.2, 9.9, -43.8]}","{'UVA': 0.1763916015625, 'UVB': 0.167266845703125, 'UVC': 0.11651611328125, 'Temperature': 27.75}",Error
-2024-07-17 17:56:13,{'Temperature': 60.05},{'Usage': 0.3},{'Usage': 12.8},Error,"{'Acceleration': [1.0414777221679687, -0.6823474731445313, -9.89523546142578], 'Gyro': [-0.012523736961832063, 0.0021316999083969466, -0.0002664624885496183], 'Magnetic': [-28.349999999999998, 10.2, -43.65]}","{'UVA': 0.07781982421875, 'UVB': 0.03460693359375, 'UVC': 0.020263671875, 'Temperature': 27.75}",Error
-2024-07-17 17:56:20,{'Temperature': 62.25},{'Usage': 0.2},{'Usage': 12.8},Error,"{'Acceleration': [1.0175357055664063, -0.603338818359375, -10.000580334472655], 'Gyro': [-0.005862174748091604, 0.00639509972519084, -0.0021316999083969466], 'Magnetic': [-28.65, 8.4, -43.05]}","{'UVA': 0.19195556640625, 'UVB': 0.14996337890625, 'UVC': 0.10638427734375, 'Temperature': 27.799999999999997}",Error
-2024-07-17 17:56:27,{'Temperature': 60.05},{'Usage': 10.0},{'Usage': 12.8},Error,"{'Acceleration': [1.0414777221679687, -0.6751648681640625, -9.97903251953125], 'Gyro': [-0.007460949679389314, 0.005062787282442748, -0.00479632479389313], 'Magnetic': [-28.05, 11.4, -43.949999999999996]}","{'UVA': 0.166015625, 'UVB': 0.132659912109375, 'UVC': 0.098785400390625, 'Temperature': 27.700000000000003}",Error
-2024-07-17 17:56:37,{'Temperature': 60.05},{'Usage': 0.3},{'Usage': 12.8},Error,"{'Acceleration': [1.0438719238281249, -0.6464344482421874, -9.9215716796875], 'Gyro': [-0.010658499541984735, 0.006928024702290077, -0.003996937328244275], 'Magnetic': [-28.05, 9.75, -43.199999999999996]}","{'UVA': 0.23345947265625, 'UVB': 0.201873779296875, 'UVC': 0.159576416015625, 'Temperature': 27.700000000000003}",Error
-2024-07-17 17:56:44,{'Temperature': 61.7},{'Usage': 0.0},{'Usage': 12.8},Error,"{'Acceleration': [1.0893617553710937, -0.6631938598632812, -9.967061511230469], 'Gyro': [-0.007727412167938932, 0.004529862305343512, -0.0002664624885496183], 'Magnetic': [-28.349999999999998, 12.15, -44.4]}","{'UVA': 0.14007568359375, 'UVB': 0.121124267578125, 'UVC': 0.078521728515625, 'Temperature': 27.75}",Error
-2024-07-17 17:56:51,{'Temperature': 59.5},{'Usage': 5.0},{'Usage': 12.8},Error,"{'Acceleration': [1.0271125122070313, -0.6703764648437499, -10.000580334472655], 'Gyro': [-0.007194487190839696, 0.003996937328244275, -0.0034640123511450386], 'Magnetic': [-28.05, 9.45, -44.4]}","{'UVA': 0.228271484375, 'UVB': 0.190338134765625, 'UVC': 0.14691162109375, 'Temperature': 27.700000000000003}",Error
-2024-07-17 17:56:57,{'Temperature': 61.7},{'Usage': 0.2},{'Usage': 12.8},Error,"{'Acceleration': [1.0917559570312498, -0.6200982299804687, -9.957484704589843], 'Gyro': [-0.007727412167938932, 0.000799387465648855, 0.002931087374045802], 'Magnetic': [-30.15, 11.1, -43.8]}","{'UVA': 0.16082763671875, 'UVB': 0.1614990234375, 'UVC': 0.101318359375, 'Temperature': 27.75}",Error
-2024-07-17 17:57:04,{'Temperature': 60.6},{'Usage': 11.0},{'Usage': 12.8},Error,"{'Acceleration': [1.0031704956054688, -0.6871358764648438, -9.986215124511718], 'Gyro': [-0.005862174748091604, 0.006661562213740459, 0.0021316999083969466], 'Magnetic': [-28.95, 9.9, -43.05]}","{'UVA': 0.1348876953125, 'UVB': 0.10382080078125, 'UVC': 0.068389892578125, 'Temperature': 27.75}",Error
-2024-07-17 17:57:11,{'Temperature': 60.6},{'Usage': 0.5},{'Usage': 12.8},Error,"{'Acceleration': [1.0678139404296874, -0.6607996582031249, -10.01015714111328], 'Gyro': [-0.007727412167938932, 0.008793262122137404, 0.000799387465648855], 'Magnetic': [-27.599999999999998, 9.45, -43.65]}","{'UVA': 0.18157958984375, 'UVB': 0.1845703125, 'UVC': 0.14691162109375, 'Temperature': 27.75}",Error
