--- Communications/CSV/18072024/data_112602.csv
@@ -1,2 +1,0 @@
-Date,CPUTemp,CPU Usage,RAM Usage,GPS,IMU,UV,Dallas Temperature
-2024-07-18 11:26:04,{'Temperature': 81.5},{'Usage': 10.1},{'Usage': 16.6},Error,"{'Acceleration': [1.0486603271484374, -0.6416460449218749, -9.943119494628906], 'Gyro': [-0.007460949679389314, 0.0061286372366412215, -0.0021316999083969466], 'Magnetic': [-8.4, 29.7, -20.849999999999998]}","{'UVA': 4.7003173828125, 'UVB': 0.19610595703125, 'UVC': 0.083587646484375, 'Temperature': 29.75}",Error
