--- Communications/CSV/18072024/data_113801.csv
@@ -1,18 +1,0 @@
-Date,CPUTemp,CPU Usage,RAM Usage,Latitude,Longitude,Altitude,Satellites,Elevation,Azimuth,UTC Time,Acceleration,Gyro,Magnetic,UVA,UVB,UVC,UV Temp,Temperature
-11:38:03,83.15,Error,16.7,40.43568266666667,-3.882774833333334,Error,Error,Error,Error,093802.00,"[0.9504980590820312, -0.7062894897460937, -9.93354268798828]","[-0.00799387465648855, 0.004529862305343512, 0.0013323124427480919]","[-7.6499999999999995, 30.45, -23.099999999999998]",4.83001708984375,0.190338134765625,0.08612060546875,30.0,Error
-11:38:10,84.25,5.3,16.8,Error,Error,Error,10,Error,Error,Error,"[0.967257470703125, -0.6775590698242188, -9.9981861328125]","[-0.00479632479389313, 0.005862174748091604, -0.0021316999083969466]","[-8.4, 27.45, -21.45]",5.00640869140625,0.282623291015625,0.149444580078125,30.0,Error
-11:38:17,81.5,14.6,16.8,Error,Error,Error,10,Error,Error,Error,"[0.9265560424804687, -0.6703764648437499, -9.947907897949218]","[-0.006928024702290077, 0.00159877493129771, 0.00319754986259542]","[-8.85, 28.5, -20.849999999999998]",4.85595703125,0.236480712890625,0.144378662109375,30.049999999999997,Error
-11:38:24,80.4,0.3,16.7,Error,Error,Error,10,Error,Error,Error,"[0.9935936889648437, -0.6679822631835937, -9.983820922851562]","[-0.0053292497709923675, 0.0005329249770992366, -0.0005329249770992366]","[-8.85, 28.05, -20.849999999999998]",4.87152099609375,0.224945068359375,0.093719482421875,30.0,Error
-11:38:34,83.7,Error,16.8,Error,Error,Error,10,Error,Error,Error,"[1.0295067138671874, -0.679953271484375, -9.909600671386718]","[-0.008526799633587786, 0.005062787282442748, -0.0018652374198473284]","[-9.0, 27.9, -21.0]",4.8248291015625,0.167266845703125,0.06585693359375,30.049999999999997,Error
-11:38:41,81.5,11.8,16.8,Error,Error,Error,10,Error,Error,Error,"[0.9576806640625, -0.689530078125, -9.986215124511718]","[-0.010924962030534353, 0.002931087374045802, 0.004529862305343512]","[-7.6499999999999995, 29.25, -21.599999999999998]",4.88189697265625,0.271087646484375,0.154510498046875,30.0,Error
-11:38:48,84.25,Error,16.6,Error,Error,Error,10,Error,Error,Error,"[1.005564697265625, -0.7110778930664062, -9.981426721191406]","[-0.005062787282442748, 0.00479632479389313, 0.003730474839694657]","[-8.85, 28.799999999999997, -22.349999999999998]",4.8455810546875,0.201873779296875,0.0911865234375,30.049999999999997,Error
-11:38:55,83.15,7.2,17.0,Error,Error,Error,10,Error,Error,Error,"[0.8930372192382812, -0.7302315063476562, -9.89523546142578]","[-0.005595712259541985, 0.0010658499541984733, 0.0018652374198473284]","[-8.25, 27.45, -20.25]",5.001220703125,0.31146240234375,0.17730712890625,30.0,Error
-11:39:01,83.7,0.2,16.7,Error,Error,Error,09,Error,Error,Error,"[0.9409212524414062, -0.6919242797851562, -9.931148486328125]","[-0.0053292497709923675, 0.002931087374045802, 0.0005329249770992366]","[-8.4, 27.75, -20.849999999999998]",4.835205078125,0.167266845703125,0.068389892578125,30.049999999999997,Error
-11:39:08,83.15,0.2,16.7,Error,Error,Error,10,Error,Error,Error,"[0.9528922607421875, -0.6679822631835937, -9.89523546142578]","[-0.00799387465648855, 0.004263399816793893, -0.002398162396946565]","[-9.299999999999999, 28.799999999999997, -20.849999999999998]",4.86114501953125,0.213409423828125,0.101318359375,30.049999999999997,Error
-11:39:15,80.4,6.2,16.7,Error,Error,Error,10,Error,Error,Error,"[1.0079588989257813, -0.699106884765625, -9.991003527832031]","[-0.008526799633587786, 0.0021316999083969466, -0.006661562213740459]","[-9.0, 29.099999999999998, -20.7]",4.918212890625,0.294158935546875,0.144378662109375,30.049999999999997,Error
-11:39:22,83.7,1.0,16.6,Error,Error,Error,10,Error,Error,Error,"[0.9624690673828125, -0.7302315063476562, -9.9981861328125]","[-0.00799387465648855, 0.002398162396946565, 0.0005329249770992366]","[-8.4, 27.3, -22.65]",4.84039306640625,0.1845703125,0.078521728515625,30.049999999999997,Error
-11:39:32,82.6,0.3,16.6,Error,Error,Error,10,Error,Error,Error,"[0.9935936889648437, -0.6631938598632812, -9.947907897949218]","[-0.008793262122137404, 0.005862174748091604, 0.00319754986259542]","[-8.7, 29.4, -21.45]",4.835205078125,0.190338134765625,0.078521728515625,30.049999999999997,Error
-11:39:39,82.6,0.2,16.6,Error,Error,Error,10,Error,Error,Error,"[0.948103857421875, -0.7158662963867187, -9.955090502929687]","[-0.006928024702290077, 0.00639509972519084, -0.000799387465648855]","[-10.049999999999999, 28.799999999999997, -20.55]",5.04791259765625,0.27685546875,0.13677978515625,30.099999999999994,Error
-11:39:46,83.7,0.5,16.6,Error,Error,Error,10,Error,Error,Error,"[0.9792284790039062, -0.7230489013671875, -9.928754284667969]","[-0.007460949679389314, 0.00639509972519084, 0.0018652374198473284]","[-9.6, 28.5, -20.099999999999998]",4.9078369140625,0.248016357421875,0.129180908203125,30.049999999999997,Error
-11:39:53,82.6,5.3,16.6,Error,Error,Error,10,Error,Error,Error,"[0.9648632690429687, -0.6775590698242188, -9.955090502929687]","[-0.005862174748091604, 0.002398162396946565, 0.0026646248854961837]","[-9.6, 28.5, -22.349999999999998]",4.959716796875,0.248016357421875,0.1317138671875,30.099999999999994,Error
-11:39:59,82.6,Error,16.7,Error,Error,Error,11,Error,Error,Error,"[0.9457096557617187, -0.6631938598632812, -9.952696301269532]","[-0.00985911207633588, 0.0034640123511450386, -0.002398162396946565]","[-10.35, 28.799999999999997, -21.15]",4.8663330078125,0.190338134765625,0.083587646484375,30.099999999999994,Error
